% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{color}
\alias{color}
\title{Color for fingerprint visulization}
\format{
A vector of 1134 character
}
\usage{
color
}
\description{
Character vector of color for fingerprint grid plot
}
\keyword{datasets}
