% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprintplot.R
\name{fingerprintplot}
\alias{fingerprintplot}
\title{This is a function for individual fingerprint  plot}
\usage{
fingerprintplot(Individual_df)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
This function will perform fingerprint heatmap in pdf file.
We assumes that you setup the specific working directory for your analysis. The result of the plot should be in your working folder.
The result of the plot should be in your working folder.
}
