% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_change.R
\name{fold_change}
\alias{fold_change}
\title{calculation of Fold-Change}
\usage{
fold_change(x)
}
\arguments{
\item{x}{Matrix of normalized expression data (not Log2 transformed).Genes should be in rows and Sample ID in columns. Row names are required to be valid Gene Symbols}
}
\value{
A matrix of the fold change comparison between "Test_group" vs ""Ref_group"
}
\description{
A function to calculate fold-change between group comparison; "Test_group" vs "Ref_group"
}
\examples{
## example sample information Example expression
## data for package testting
Test_sample = matrix(data = rexp(1000, rate = 0.01),
                    nrow = 14168, ncol = 20)
control_sample = matrix(data = rexp(1000, rate = 0.1),
                       nrow = 14168, ncol = 10)
data.matrix = data.frame(cbind(Test_sample, control_sample))
data.matrix$Symbol = Module_listGen3$Gene
data.matrix = aggregate(data.matrix[,-31], FUN = mean, by = list(data.matrix$Symbol))
rownames(data.matrix) = data.matrix$Group.1
data.matrix$Group.1 = NULL
colnames(data.matrix) = c(paste0(rep("SampleID", 30),
                                1:30))
## Example of ample information
sample_ann = data.frame(SampleID = (colnames(data.matrix)),
                       Group_test = c(rep("Test", 20), rep("Control",
                                                           10)), stringsAsFactors = FALSE)
rownames(sample_ann) = sample_ann$SampleID

FCgroup = fold_change(data.matrix)

}
\author{
Darawan Rinchai \href{mailto:drinchai@gmail.com}{drinchai@gmail.com}
}
