% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Gen3_ann}
\alias{Gen3_ann}
\title{Functional annotation for visulization}
\format{
A data.frame with 6 variables:\code{Modules}, \code{Aggregate=Cluster}, \code{Sub_aggregate=Cluster_location},\code{Function},\code{Position},\code{Module_color}
}
\usage{
Gen3_ann
}
\description{
Functional annotation for visulization
}
\keyword{datasets}
