% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridplot.R
\name{gridplot}
\alias{gridplot}
\title{This is a function for fingerprint grid plot}
\usage{
gridplot(Group_df)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
This function will perform grid plot in pdf file.
We assumes that you setup the specific working directory for your analysis. The result of the plot should be in your working folder.
The result of the plot should be in your working folder.
}
