% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Individualcomparison.R
\name{Individualcomparison}
\alias{Individualcomparison}
\title{Module individual fingerprint analysis}
\usage{
Individualcomparison(data.matrix)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
This function will perform single sample analysis.
We assumes that your expression file = data.matrix and annotation file = sample.info.
In the sample_test table must contains a colum name "Group_test" with specific information which sample ID = test or control
}
