% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Module_listGen3}
\alias{Module_listGen3}
\title{Module identifer and list membership in each module}
\format{
A data.frame with 14168 rows by 5 variables:
\code{Module: Module identifier},
\code{Gene symbol: gene membership},
\code{Module_gene: gene specific module membership},
\code{Function: Functional annotation},
\code{position: position on fingerprint grid plot}
}
\usage{
Module_listGen3
}
\description{
A dataframe contain gene member of 3rd generantion of blood module repertoire construction
}
\keyword{datasets}
