% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprintplot.R
\name{fingerprintplot}
\alias{fingerprintplot}
\title{Individual fingerprint visualization}
\usage{
fingerprintplot(
  Individual_df,
  cutoff = NULL,
  rowSplit = TRUE,
  Ref_group = NULL,
  Group_column = NULL,
  Aggregate = NULL,
  height = NULL,
  width = NULL
)
}
\arguments{
\item{Individual_df}{Output table generated after running the 'Individualcomparison' function}

\item{cutoff}{Sets the percentage cut off used for fingerprint visualization, range of acceptable values from 0 to 100}

\item{rowSplit}{Splits row of heatmap by each aggregate}

\item{Ref_group}{Reference group or samples that considered as control}

\item{Group_column}{Name of the columns for the groups used for the analysis}

\item{Aggregate}{Selects specific module aggregates for heatmap fingerprint plot}

\item{height}{Sets height dimension for the heatmap plot}

\item{width}{Sets width dimension for the heatmap plot}
}
\value{
A heatmap of \% of module response in each single sample
}
\description{
The fingerprintplot function will generate fingerprint heatmap plots as a pdf file. The file will be saved in the working directory specified for the analysis.
The default cut off for visualization is set at 15\%, it can changed to any value between 0-100\%.
}
