% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprintplot.R
\name{fingerprintplot}
\alias{fingerprintplot}
\title{Individual fingerprint visualization
The fingerprintplot function will generate fingerprint heatmap plots as a pdf file. The file will be saved in the working directory specified for the analysis.
The default cut off for visualization is set at 15\%, it can be changed to any value between 0-100\%.}
\usage{
fingerprintplot(
  Individual_df,
  sample_info = sample_info,
  cutoff = NULL,
  rowSplit = TRUE,
  Ref_group = NULL,
  show_ref_group = FALSE,
  Group_column = NULL,
  Aggregate = NULL,
  filename = NULL,
  height = NULL,
  width = NULL
)
}
\arguments{
\item{Individual_df}{Dataframe with output generated after running the 'Individualcomparison' function}

\item{sample_info}{A dataframe with sample annotation. Sample_info dataframe requires two columns: Sample ID (exactly matching Sample ID of         data.matrix) and a column specifying group annotation}

\item{cutoff}{SNumeric value specifying the percentage cut off used for fingerprint visualization (range of acceptable values from 0 to 100).}

\item{rowSplit}{Logical operator (TRUE/FALSE) to indicate if rows of the heatmaps should be split by each aggregate}

\item{Ref_group}{Characters name of reference group or samples that considered as control (Example: Control, baseline, Pre-treatment,... etc)}

\item{show_ref_group}{Character vector specifying value within the group column         (Group_column) that will be used as Reference group (samples considered as control)}

\item{Group_column}{Name of the columns for the groups used for the analysis}

\item{Aggregate}{Character vector specifying name of specific module aggregates    for heatmap fingerprint plot}

\item{filename}{Character vector with a name for saving file}

\item{height}{Sets the height of the graphics region in inches. The default values are 28}

\item{width}{Sets the width of the graphics region in inches. The default values are 17}
}
\value{
\preformatted{          A heatmap of \% of module response in each single sample
}
}
\description{
Individual fingerprint visualization
The fingerprintplot function will generate fingerprint heatmap plots as a pdf file. The file will be saved in the working directory specified for the analysis.
The default cut off for visualization is set at 15\%, it can be changed to any value between 0-100\%.
}
\author{
Darawan Rinchai \href{mailto:drinchai@gmail.com}{drinchai@gmail.com}
}
