% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Gen3_ann}
\alias{Gen3_ann}
\title{Functional annotation for visulization}
\format{
A data.frame with 5 variables:
\code{Modules: Module identifier},
\code{Function: Functional annotation},
\code{Position: position on fingerprint grid plot},
\code{Module_color: specific color of each module for visulization},
\code{Cluster: Module cluster membership}
}
\usage{
Gen3_ann
}
\description{
A dataframe contain module identifer, functional annotation and color in each specific module
}
\keyword{datasets}
