% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridplot.R
\name{gridplot}
\alias{gridplot}
\title{##Fingerprint grid visualization}
\usage{
gridplot(Group_df, cutoff = NULL, Ref_group = NULL, filename = NULL)
}
\arguments{
\item{Group_df}{Output matrix generated after running the 'Groupcomparison' function}

\item{cutoff}{Sets the percentage cut off used for fingerprint visualization, range of acceptable values from 0 to 100}

\item{Ref_group}{Characters name of reference group or samples that considered as control (Example: Control, baseline, Pre-treatment,... etc)}

\item{filename}{Give a file name for grid plot}
}
\value{
\preformatted{       A pdf file of grid plot
}
}
\description{
##The gridplot function will generate a grid plot as a pdf file. Specific working directory for the analysis need to be specified for saving the file. The result of the plot should be return in the same working directory.
##The default cut off for visualization is set at 15\%, it can be changed to any value between 0-100\%.
}
\examples{
gridplot(Group_df, cutoff = 15, Ref_group = "Control",filename="Group_comparison_cutoff15")
#' @author
Darawan Rinchai <drinchai@gmail.com>
}
