% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridplotlimma.R
\name{gridplotlimma}
\alias{gridplotlimma}
\title{Fingerprint grid visualization after performing "Groupcomparisonlimma"}
\usage{
gridplotlimma(
  data = Group_limma,
  cutoff = NULL,
  Ref_group = NULL,
  filename = NULL
)
}
\arguments{
\item{cutoff}{Sets the percentage cut off used for fingerprint visualization, range of acceptable values from 0 to 100}

\item{Ref_group}{Reference group or samples that considered as control}

\item{filename}{Give a file name for grid plot}

\item{Group_limma}{Output table generated after running the 'Groupcomparison.limma' function}
}
\value{
\preformatted{          A pdf file of grid plot
}
}
\description{
The gridplot function will generate a grid plot as a pdf file. Specific working directory for the analysis need to be specified for saving the file. The result of the plot should be return in the same working directory.
The default cut off for visualization is set at 15\%, it can be changed to any value between 0-100\%.
}
\examples{
gridplot(Group_df, cutoff = 15, Ref_group = "Control",filename="Limma_group_comparison")
}
\author{
Darawan Rinchai \href{mailto:drinchai@gmail.com}{drinchai@gmail.com}
} 
