% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Groupcomparison.R
\name{Groupcomparison}
\alias{Groupcomparison}
\title{Group comparison analysis}
\usage{
Groupcomparison(
  data.matrix,
  sample_info = sample_info,
  FC = NULL,
  pval = NULL,
  FDR = TRUE,
  Group_column = NULL,
  Ref_group = NULL
)
}
\arguments{
\item{data.matrix}{A table of normalized expression matrix (Important: Expression matrix must be none Log2 transformed as it will be automatic transformed when running this function)}

\item{sample_info}{A table of sample information (Important: rownames of sample information must be the same names as in colnames of data.matrix)}

\item{FC}{Foldchange cut off to consider th eabundance of a given transcript to be increased or decreased compared to a reference group (Ref_group)}

\item{pval}{p-value cut off or False discovery rate when FDR = FALSE}

\item{FDR}{False discovery rate cut off using BH method}

\item{Group_column}{Name of the columns for the groups used for the analysis}

\item{Ref_group}{Characters name of reference group or samples that considered as control (Example: Control, baseline, Pre-treatment,... etc)}
}
\value{
\preformatted{            A matrix of the percentahe of module response in each group comparison
}
}
\description{
The Groupcomparison function will perform group comparison analyses and the results are expressed “at the module level” as percent of genes increased or decreased.
\itemize{
\item Expression matrix and sample annotation files are required to perform this analysis.
\item The sample annotation file must be loaded using a specific name = "sample_info".
\item The names of the columns for the conditions used in the analysis must be specified.
}
}
\examples{
Group_df <- Groupcomparison(data.matrix, sample_info = sample_info, FC = 0, pval = 0.1, FDR = TRUE, Group_column = "Group_test", Ref_group = "Control")
}
\author{
Darawan Rinchai \href{mailto:drinchai@gmail.com}{drinchai@gmail.com}
}
