% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Groupcomparison.R
\name{Groupcomparison}
\alias{Groupcomparison}
\title{Module group comparison analysis}
\usage{
Groupcomparison(data.matrix)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
This function will perform group comparison analysis.
We assumes that your expression file = data.matrix and annotation file = sample.info.
#' In the sample_test table must contains a colum name "Group_test" with specific information which sample ID = group1, group2 or control
}
