% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Groupcomparisonlimma.R
\name{Groupcomparisonlimma}
\alias{Groupcomparisonlimma}
\title{Group comparison analysis using limma}
\usage{
Groupcomparisonlimma(
  data = data.matrix,
  sample_info = sample_info,
  FC = NULL,
  pval = NULL,
  FDR = TRUE,
  Group_column = NULL,
  Test_group = "Test_group",
  Ref_group = NULL
)
}
\arguments{
\item{sample_info}{A table of sample information (Important: rownames of sample information must be the same names as in colnames of data.matrix)}

\item{FC}{Foldchange cut off to consider th eabundance of a given transcript to be increased or decreased compared to a reference group (Ref_group)}

\item{pval}{p-value cut off or False discovery rate when FDR = FALSE}

\item{FDR}{False discovery rate cut off using BH method}

\item{Group_column}{Name of the columns for the groups used for the analysis}

\item{Test_group}{Test group or samples that considered as test}

\item{Ref_group}{Reference group or samples that considered as control}

\item{data.matrix}{Normalized expression data (Important: Expression matrix must be none Log2 transformed as it will be automatic transformed when running this function)}
}
\value{
\preformatted{         A matrix of the percentahe of module response in each group comparison
}
}
\description{
The Groupcomparisonlimma function will perform group comparison analyses using "limma" function from "limma R package" and the results are expressed “at the module level” as percent of genes increased or decreased.
\itemize{
\item Expression matrix and sample annotation files are required to perform this analysis.
\item The sample annotation file must be loaded using a specific name = "sample_info".
\item The names of the columns for the conditions used in the analysis must be specified.
}
}
\examples{
Group_df <- Groupcomparisonlimma(data.matrix,sample_info = sample_info, FC = 0, pval = 0.1, FDR = TRUE, Group_column = "Group_test", Test_group = "Sepsis", Ref_group = "Control")
}
\author{
Darawan Rinchai \href{mailto:drinchai@gmail.com}{drinchai@gmail.com}
}
