% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Individualcomparison.R
\name{Individualcomparison}
\alias{Individualcomparison}
\title{Individual single sample analysis}
\usage{
Individualcomparison(
  data.matrix,
  FC = NULL,
  DIFF = NULL,
  Group_column = NULL,
  Ref_group = NULL
)
}
\arguments{
\item{data.matrix}{Raw normalized expression data}

\item{FC}{foldchange cut off to consider increasing or decreasing as comparison to reference sample group (Ref_group)}

\item{DIFF}{difference cut off to consider increasing or decreasing as comparison to reference sample group (Ref_group)}

\item{Group_column}{Column name of the conditions that considered for analysis}

\item{Ref_group}{Reference group or samples that considered as control}
}
\value{
A matrix of the percentahe of module response at individual level
}
\description{
Individualcomparison function will perform single sample analysis as comparison to control and the results are expressed “at the module level” as percent of genes increased or decreased.
Annotation of the gene expression data matrix with module membership information. Then differential expression and calculation of percentage response
Expression matrix and sample annotation file is required for performing this analysis. The sample annotation file must loaded as specific name = "sample.info"
A specific column of conditions for analysis must be indicated
}
\examples{
Individual_df = Individualcomparison(data.matrix, FC = 1.5, DIFF = 10, Group_column = "Group_test",Ref_group = "Control")
}
\author{
Darawan Rinchai \href{mailto:drinchai@gmail.com}{drinchai@gmail.com}
}
