% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Individualcomparison.R
\name{Individualcomparison}
\alias{Individualcomparison}
\title{Individual single sample analysis}
\usage{
Individualcomparison(
  data.matrix,
  sample_info = sample_info,
  FC = NULL,
  DIFF = NULL,
  Group_column = NULL,
  Ref_group = NULL
)
}
\arguments{
\item{data.matrix}{Normalized expression data (Important: Expression matrix must be none Log2 transformed as it will be automatic transformed when running this function)}

\item{sample_info}{A table of sample information (Important: rownames of sample information must be the same names as in colnames of data.matrix)}

\item{FC}{Foldchange cut off to consider th eabundance of a given transcript to be increased or decreased compared to a reference group (Ref_group)}

\item{DIFF}{Difference cut off to consider th eabundance of a given transcript to be increased or decreased compared to a reference group (Ref_group)}

\item{Group_column}{Name of the columns for the groups used for the analysis}

\item{Ref_group}{Reference group or samples that considered as control}
}
\value{
\preformatted{         A matrix of the percentahe of module response at individual level
}
}
\description{
The Individualcomparison function will perform individual sample comparison analysis in reference to a control sample or group of samples, with the results are expressed “at the module level” as percent of genes increased or decreased.
}
\details{
\itemize{
\item Expression matrix and sample annotation file are required in order to perform this analysis.
\item The sample annotation file must be loaded using a specific name = "sample_info".
\item The names of the columns for the conditions used in the analysis must be specified
\item The default cutoff is set at FC =1.5 and DIFF =10
}
}
\examples{
Individual_df = Individualcomparison(data.matrix, sample_info = sample_info,FC = 1.5, DIFF = 10, Group_column = "Group_test",Ref_group = "Control")
}
\author{
Darawan Rinchai \href{mailto:drinchai@gmail.com}{drinchai@gmail.com}
}

