% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Module_listGen3}
\alias{Module_listGen3}
\title{Module list member}
\format{
A data.frame with 14168 rows by 8 variables:\code{Module}, \code{Gene symbol}, \code{Probe ID}, \code{Module_gene},\code{Functiom},\code{Cluster},\code{Cluster_location},\code{position}
}
\usage{
Module_listGen3
}
\description{
A dataframe contain member of 3rd generantion of blood module repertoire construction
}
\keyword{datasets}
