% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Module_listGen3}
\alias{Module_listGen3}
\title{Module list member}
\format{
A 14168 by 8 data.frame
}
\usage{
Module_listGen3
}
\description{
A dataframe contain member of 3rd generantion of blood module repertoire construction
}
\keyword{datasets}
